+++ src/proto/channel.pro
@@ -3,13 +3,17 @@ void ch_logfile(char_u *fname, char_u *opt);
 int ch_log_active(void);
 channel_T *add_channel(void);
 int has_any_channel(void);
+int channel_still_useful(channel_T *channel);
+int channel_can_close(channel_T *channel);
 int channel_unref(channel_T *channel);
 int free_unused_channels_contents(int copyID, int mask);
 void free_unused_channels(int copyID, int mask);
 void channel_gui_register_all(void);
 channel_T *channel_open(const char *hostname, int port, int waittime, void (*nb_close_cb)(void));
+void ch_close_part(channel_T *channel, ch_part_T part);
 void channel_set_pipes(channel_T *channel, sock_T in, sock_T out, sock_T err);
 void channel_set_job(channel_T *channel, job_T *job, jobopt_T *options);
+void channel_write_in(channel_T *channel);
 void channel_buffer_free(buf_T *buf);
 void channel_write_any_lines(void);
 void channel_write_new_lines(buf_T *buf);
@@ -36,30 +40,6 @@ int channel_select_check(int ret_in, void *rfds_in, void *wfds_in);
 int channel_parse_messages(void);
 int channel_any_readahead(void);
 int set_ref_in_channel(int copyID);
-void clear_job_options(jobopt_T *opt);
-int get_job_options(typval_T *tv, jobopt_T *opt, int supported, int supported2);
-void job_free_all(void);
-int job_any_running(void);
-int win32_build_cmd(list_T *l, garray_T *gap);
-void job_cleanup(job_T *job);
-int set_ref_in_job(int copyID);
-void job_unref(job_T *job);
-int free_unused_jobs_contents(int copyID, int mask);
-void free_unused_jobs(int copyID, int mask);
-job_T *job_alloc(void);
-void job_set_options(job_T *job, jobopt_T *opt);
-void job_stop_on_exit(void);
-int has_pending_job(void);
-int job_check_ended(void);
-job_T *job_start(typval_T *argvars, char **argv_arg, jobopt_T *opt_arg, job_T **term_job);
-char *job_status(job_T *job);
-int job_stop(job_T *job, typval_T *argvars, char *type);
-void invoke_prompt_callback(void);
-int invoke_prompt_interrupt(void);
-void f_prompt_getprompt(typval_T *argvars, typval_T *rettv);
-void f_prompt_setcallback(typval_T *argvars, typval_T *rettv);
-void f_prompt_setinterrupt(typval_T *argvars, typval_T *rettv);
-void f_prompt_setprompt(typval_T *argvars, typval_T *rettv);
 void f_ch_canread(typval_T *argvars, typval_T *rettv);
 void f_ch_close(typval_T *argvars, typval_T *rettv);
 void f_ch_close_in(typval_T *argvars, typval_T *rettv);
@@ -78,10 +58,4 @@ void f_ch_evalraw(typval_T *argvars, typval_T *rettv);
 void f_ch_sendraw(typval_T *argvars, typval_T *rettv);
 void f_ch_setoptions(typval_T *argvars, typval_T *rettv);
 void f_ch_status(typval_T *argvars, typval_T *rettv);
-void f_job_getchannel(typval_T *argvars, typval_T *rettv);
-void f_job_info(typval_T *argvars, typval_T *rettv);
-void f_job_setoptions(typval_T *argvars, typval_T *rettv);
-void f_job_start(typval_T *argvars, typval_T *rettv);
-void f_job_status(typval_T *argvars, typval_T *rettv);
-void f_job_stop(typval_T *argvars, typval_T *rettv);
 /* vim: set ft=c : */
