+++ src/proto/edit.pro
@@ -4,9 +4,7 @@ int ins_need_undo_get(void);
 void ins_redraw(int ready);
 int decodeModifyOtherKeys(int c);
 void edit_putchar(int c, int highlight);
-char_u *buf_prompt_text(buf_T* buf);
-char_u *prompt_text(void);
-int prompt_curpos_editable(void);
+void set_insstart(linenr_T lnum, int col);
 void edit_unputchar(void);
 void display_dollar(colnr_T col);
 void undisplay_dollar(void);
